+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.3.0
+
+* Added a new filter for quests with low time remaining. 6 hours currently, will add a config option later.
+* Added an icon that displays when quests are low on time. 75 minutes currently, matching the default UI.
+
 ## v0.2.1
 
 * Change tracking interactions to match quest log/map icons